pyproject.toml
[project]
name = "artrefsync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "dataclasses-json>=0.6.7",
    "fast-autocomplete>=0.9.0",
    "levenshtein>=0.27.3",
    "lxml>=6.0.2",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "ruff>=0.14.7",
    "setuptools>=80.9.0",
    "simple-toml-configurator>=1.3.0",
    "ttkbootstrap>=1.19.2",
]
